+++ modules/constants.lua
@@ -1,6 +1,9 @@
 local constants = {}
 
-constants.mercuryVersion = "1.0.6"
+constants.mercuryVersion = "1.1.0"
 constants.xd3CmdLine = "xdelta3 -d -s \"%s\" \"%s\" \"%s\""
+constants.latestReleaseApi =
+    "https://api.github.com/repos/Sledmine/Mercury/releases/latest"
+constants.gitHubReleases = "https://github.com/Sledmine/Mercury/releases/{tagName}/"
 
 return constants